// src/app/models/articles.ts
export interface Articles {
    id?: number;
    title: string;
    source?: string;
    status: string;
    date: Date;
}

// export const ARTICLES: Articles[] = [
//     {id: 1, title: 'Labas', source: 'delfi', status: 'active'},
//     {id: 2, title: 'Labas', source: 'bbc', status: 'active'},
//     {id: 3, title: 'Labas', status: 'active'},
//     {id: 4, title: 'Labas', source: '15min', status: 'active'},
//     {id: 5, title: 'Labas', source: 'financialtimes', status: 'active'},
// ]